restructure src
